requirements.txt
web3==6.0.0
python-dotenv==1.0.0
requests==2.31.0
tabulate==0.9.0
aiohttp==3.8.5
asyncio==3.4.3
eth-account==0.9.0
matplotlib==3.7.2
pandas==2.0.3
loguru==0.7.2
python-telegram-bot==20.4 
schedule==1.2.0 